requirements.txt
python-dotenv
pinecone-client
openai==0.27.1
sentence-transformers
